requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.0
altgraph==0.17.4
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
-e git+https://github.com/huggingface/diffusers@5a6922786322fb45bd011dd6389e30e7adcdc947#egg=diffusers
docker-pycreds==0.4.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fiona==1.9.6
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
grad-cam==1.5.0
grpcio==1.62.2
h5py==3.10.0
huggingface-hub==0.24.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.19.0
jax==0.4.23
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonargparse==4.27.7
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
lightly==1.5.1
lpips==0.1.4
lightly-utils==0.0.2
lightning==2.2.1
lightning-utilities==0.10.1
macholib==1.16.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
ogb==1.2.6
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
outdated==0.2.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
peft==0.7.0
pexpect==4.9.0
Pillow==10.2.0
platformdirs==4.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.14
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
rasterio==1.3.9
regex==2024.5.15
requests==2.31.0
rich==13.7.1
Rtree==1.2.0
safetensors==0.4.2
seaborn==0.13.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.1
segmentation-models-pytorch==0.4.0
sentry-sdk==1.40.6
setproctitle==1.3.3
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.8
tokenizers==0.15.2
torch==2.4.1
torch-fidelity==0.3.0
torch-geometric==2.0.1
torchgeo==0.5.2
torchmetrics==1.3.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
ttach==0.0.3
typeshed_client==2.5.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.42.0
wilds==2.0.0
yarl==1.9.4
zennit==0.5.1
zipp==3.18.1
xformers==0.0.28
absl-py==2.1.0
accelerate==0.29.3
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.0
altgraph==0.17.4
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
-e git+https://github.com/huggingface/diffusers@5a6922786322fb45bd011dd6389e30e7adcdc947#egg=diffusers
docker-pycreds==0.4.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fiona==1.9.6
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
grad-cam==1.5.0
grpcio==1.62.2
h5py==3.10.0
huggingface-hub==0.24.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.19.0
jax==0.4.23
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonargparse==4.27.7
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
lightly==1.5.1
lpips==0.1.4
lightly-utils==0.0.2
lightning==2.2.1
lightning-utilities==0.10.1
macholib==1.16.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
ogb==1.2.6
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
outdated==0.2.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
peft==0.7.0
pexpect==4.9.0
Pillow==10.2.0
platformdirs==4.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.14
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
rasterio==1.3.9
regex==2024.5.15
requests==2.31.0
rich==13.7.1
Rtree==1.2.0
safetensors==0.4.2
seaborn==0.13.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.1
segmentation-models-pytorch==0.4.0
sentry-sdk==1.40.6
setproctitle==1.3.3
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.8
tokenizers==0.15.2
torch==2.4.1
torch-fidelity==0.3.0
torch-geometric==2.0.1
torchgeo==0.5.2
torchmetrics==1.3.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
ttach==0.0.3
typeshed_client==2.5.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.42.0
wilds==2.0.0
yarl==1.9.4
zennit==0.5.1
zipp==3.18.1
xformers==0.0.28

